+++ lib/generators/active_record/templates/migration.rb
@@ -1,18 +1,5 @@
 class RygsaekCreate<%= table_name.camelize %> < ActiveRecord::Migration
   def change
-    create_table(:<%= table_name %>) do |t|
-<%= migration_data -%>
-
-<% attributes.each do |attribute| -%>
-      t.<%= attribute.type %> :<%= attribute.name %>
-<% end -%>
-
-      t.timestamps
-    end
-
-    # add_index :<%= table_name %>, :email,                unique: true
-    # add_index :<%= table_name %>, :reset_password_token, unique: true
-    # add_index :<%= table_name %>, :confirmation_token,   unique: true
-    # add_index :<%= table_name %>, :unlock_token,         unique: true
+    <%= migration_data -%>
   end
 end