+++ lib/generators/rygsaek/orm_helpers.rb
@@ -34,6 +34,11 @@ CONTENT
       def model_exists?
         File.exists?(File.join(destination_root, model_path))
       end
+      
+      def rygsaek_tables_exists?
+        puts file_path
+        File.exists?(File.join(destination_root, model_path))
+      end
 
       def migration_exists?(table_name)
         Dir.glob("#{File.join(destination_root, migration_path)}/[0-9]*_*.rb").grep(/\d+_add_rygsaek_to_#{table_name}.rb$/).first