+++ test/generators/rygsaek_generator_test.rb
@@ -22,11 +22,11 @@ class RygsaekGeneratorTest < Rails::Generators::TestCase
     assert_file "config/routes.rb", match
   end
 
-  test "route generation with skip routes" do
-    run_generator %w(monster name:string --skip-routes)
-    match = /rygsaek_on :monsters, skip: :all/
-    assert_file "config/routes.rb", match
-  end
+  # test "route generation with skip routes" do
+  #   run_generator %w(monster name:string --skip-routes)
+  #   match = /rygsaek_on :monsters, skip: :all/
+  #   assert_file "config/routes.rb", match
+  # end
 
   def copy_routes
     routes = File.expand_path("../../rails_app/config/routes.rb", __FILE__)